refactor(record-list): remove duplicate code
