+++ apps/metadata-editor/src/app/dashboard/search-header/search-header.component.spec.ts
@@ -2,28 +2,15 @@ import { ChangeDetectionStrategy, CUSTOM_ELEMENTS_SCHEMA } from '@angular/core'
 import { ComponentFixture, TestBed } from '@angular/core/testing'
 
 import { SearchHeaderComponent } from './search-header.component'
-import { BehaviorSubject, of } from 'rxjs'
-import { barbieUserFixture } from '@geonetwork-ui/common/fixtures'
+import { of } from 'rxjs'
 import { StoreModule } from '@ngrx/store'
 import { EffectsModule } from '@ngrx/effects'
 import { TranslateModule } from '@ngx-translate/core'
 import { PlatformServiceInterface } from '@geonetwork-ui/common/domain/platform.service.interface'
 import { AvatarServiceInterface } from '@geonetwork-ui/api/repository'
 import { SearchService } from '@geonetwork-ui/feature/search'
-
-class AvatarServiceInterfaceMock {
-  getPlaceholder = () => of('http://placeholder.com')
-  getProfileIcon = (hash: string) => of(`${hash}`)
-}
-
-const me$ = new BehaviorSubject(barbieUserFixture())
-class PlatformServiceMock {
-  getMe = jest.fn(() => me$)
-}
-
-class SearchServiceMock {
-  updateFilters = jest.fn()
-}
+import { MockProvider, MockProviders } from 'ng-mocks'
+import { RouterFacade } from '@geonetwork-ui/feature/router'
 
 describe('SearchHeaderComponent', () => {
   let component: SearchHeaderComponent
@@ -38,18 +25,14 @@ describe('SearchHeaderComponent', () => {
         TranslateModule.forRoot(),
       ],
       providers: [
-        {
-          provide: AvatarServiceInterface,
-          useClass: AvatarServiceInterfaceMock,
-        },
-        {
-          provide: PlatformServiceInterface,
-          useClass: PlatformServiceMock,
-        },
-        {
-          provide: SearchService,
-          useClass: SearchServiceMock,
-        },
+        MockProviders(
+          AvatarServiceInterface,
+          PlatformServiceInterface,
+          SearchService
+        ),
+        MockProvider(RouterFacade, {
+          currentRoute$: of(null),
+        }),
       ],
     })
       .overrideComponent(SearchHeaderComponent, {
